Refs #755 - Implement monitoring of player online status
Fixed issue with double notification. The Observable() was raising a
'recursive call' error, due to the fact that we had two different
notifications happening at the same time.

Modified the lock mechanism to allow to allow to have several
notifications at the same time from the same object, as long as they are
from different types. Added the functionality in a separate Lockable()
object to allow reuse.