Friendlier panic error messages (closes #808).
I wanted to display the panic message not as an alert box,
but using our show_warning machinery which is already in place.
I was thinking that if using rich HTML, I would still be able to
display details about the error, that would be hidden by default
and only shown if the user clicked on a link somewhere.

Using show_warning proved to be rather tricky, since show_warning
needs to get passed player_id, game_id, and the root in order to
properly set up the board buttons and background polls.
This was the reason I had to modify quite some functions,
among them $.cardstories.ajax, $.cardstories.poll, $.cardstories.send
to take these extra parameters.

Since it took quite some work, I didn't get to implement my initial
idea that there would be a way to get to more detailed information.
It could still be implemented one day if it proves it would be useful.