Tallying of player scores (see #769)
This implements the tallying of player scores upon completion of a game.
All data is stored on the WS database, as implemented in #767. The
algorythm itself is described in #817.  As of yet there is no visible
indication of scores - the frontend will be developed in a future task.

Of note:

* The actual point values are defined as constants.
* Points are *only* tallied upon game completion, and nowhere else.
* A small (but essential) database change was necessary, but due to the
  minor nature of the change itself and the fact that the model in
  question was not in use until now, the previous migration file was
  updated, instead of adding a new one.