Create game before page is loaded (closes #829).
A new state called 'create' has been added to the game states.
This is the state that every new game is initially in.
When the game is created, the only thing that is known about
it is the owner.
The owner picks the card and sets the sentence in two separate
steps, called 'set_card' and 'set_sentence' on the service.

There was previously a function on the JS client called 'create',
which was renamed to 'create_new_game' and is responsible for
sending the initial request to create a new game to the service.
The new $.cardstories.create function is used to handle display
of the games in 'create' state, based on the same convention
as $.cardstories.invitation, $.cardstories.vote, and
$.cardstories.complete.

If a player other than the owner tries to connect to a game
that isn't ready to be played yet (the card and/or sentence)
haven't been set yet, the $.cardstories.create_wait function
is called, which now simply alerts the information that game
isn't ready yet. I didn't try to do anything more fancy than
that at this stage, as I imagine tables will want to handle
this state, so I left this for another task.

This commit includes a migration (number 003), which changes
the default value of games.state field to 'create'.
Since SQLite doesn't have a 'modify column' command, the whole
table needs to be copied to a temporary table and then recreated.
Creating a backup of the database before running the migration
is strongly recommended.

One of the table server plugin tests is failing. I believe this
to be the problem with the test itself. I spent some time on it,
trying to fix it, but since I didn't seem to be coming any closer
to a solution, I gave up on it as tables are being rewritten anyway
and the test probably soon won't be relevant anymore.

As a minor change I also removed the check in CardstoriesException
for whether the error code is one of the known error codes.
That check was supposed to make it easier to keep error codes in
sync with the client, but has proved to be more trouble than it was
worth, so I decided to remove it.