Make players see what GM is doing in 'create' state (closes #838).
Players can now join games in 'create' state. Player can see his
seat when joining a game in 'create' state, as well as other joined
players' seats (the author doesn't see players joining yet, this is
to be done in a separate task).

The "waiting" for game to be created was split into three parts:
waiting for the card, waiting for the sentence, and transitioning
from 'create' into 'invitation' state.

Most of the logic and animations have been reused from the 'invitation'
state and code modified accordingly. The 'cardstories_invitation' HTML
div is reused for this state, too, which goes against the principle
that every state should have its own HTML div, but I didn't have the
courage to just copy&paste 'cardstories_invitation' into a new div :).

While waiting for the GM to pick a card, a dialog box is shown informing
the player what he is waiting for. While waiting for the GM to write
the story, a similar dialog is shown.
The text in the dialog might have to be improved as it's pretty minimal
right now.

Finally, I added appropriate skin states, and slightly modified the
look of the skin links as they were getting hard to navigate around.