Implements server-side levels (fixes #853)
This implements the level-up logic as specified in #648.  The three
coefficients are tunable by editing the appropriate constants
(LEVEL_[A|B|C]) in game.py.

Note: in writing the tests for this, I noticed that on several other
spots on test_game.py a poll is set up for post-complete tests.
However, it seems that calling self.assert* from inside these callbacks
has no effect on test results.  This needs to be corrected in the
future.