Animate score numbers (fixes #860)
This animates the score numbers from the previous value to the current
one (set upon completion).

Note: the $.cardstories.complete() function now returns a deferred, so
that it can be properly tested after completion of all animations.  This
was needed as the score animation uses timeouts to achieve the desired
effect.