Make the author see players joining in 'create' state (closes #855).
The author see players as they are joining the game in 'create' state.
The player seat is gently faded in as players arrive. The tricky part
was not replaying the join animations for all players when switching
state.

Because when switching from pick_card -> write_sentence -> invitation_owner
states we are switching HTML div elements, the 'noop' CSS class trick doesn't
always work, therefore a more persistent state was needed.
I used the $root.data('cardstories_state') state (which includes a copy of
the game state as it was before the last set_active call was made).

The cleaning up of the skin.js file was just me getting frustrated
by some issues I was hitting while implementing #855, so I took a
couple of minutes off and relaxed while going trough skin.js.