Fixes #868. Requested player_id/owner_id = undefined
This fixes a problem that occurred right after a player logged in using
django: a second 'statechange' event with empty data would be caught by
$.cardstories.onstatechange() right after the first one, and load_game
would try to proceed without a player_id (an illegal situation).  One of
the first side effects would be the server-side exception upon
authentication.

I have not ascertained exactly *why* this second statechange occurs, but
I've opted to prepare for it in the event handler - if a non-pushState()
originated state change occurs, just do nothing.