Partially fix broken tables (refs #865).
The automatic redirecting players to a new game once the current one
has finished should be working, however I deactiveated the redirect
where a user first loads the page and the table plugin tried to redirect
him/her to an available table.
The way it was implemented didn't work well with the new loading system
(without page reloading), so I decided to deactivate it for now and tackle
it in a separate step.