Fix the table plugin (closes #865).
This is the second (and hopefully final) commit for bug #865.

This one takes care of redirecting the player to an available table
when he first loads the page. This is now done as part of the bootstrap
code in $.cardstories, rather than as part of the $.cardstories_table
plugin, as I couldn't come up with any other way that wouldn't be too hackish.

The sole task of finding a suitable game is still delegated to the table plugin
throught the $.cardstories_table.get_available_game function.