Fix broken table plugin test (closes #878).
This turned out to be problem with actual code, not just a problem with
the test as I had myself believe.
There was a check in the code that, when a new game was created, would ONLY
touch/return the poll if there previouly wasn't any available game.

The code was broken (twice) since "yield not something()" doesn't really work
as you'd expect (it always returns False, regardless of what the call to something returns);
and because it was checking for presence of an available table right AFTER a new
table was constructed. As seen from the logs there was some code shuffling going on
in this area and I guess this code fragment ended in a wrong position by accident.

But since I don't really think it matters whether the poll returns on every new game
(I don't believe we were relying on that anywhere, though I may be wrong), I decided
to remove that check altogether for simplicity.