Upgrades to jQuery 1.7.1 (see #877)
This upgrades us to jQuery 1.7.1, primarily to get rid of a deprecation
warning in Chrome 16+.  QUnit was also upgrade to 1.4.0.

Relatively few changes to the code were necessary, and these were only
needed to get unit tests working.

The game itself was also tested by hand from start to finish several
times on different browsers, with the notable exception of Safari on
OSX.