Version critical CSS and JS files (fixes #86)
This implements a simple way to version our critical JS and CSS files:
it adds a query parameter to their URLs that contains their last
modification time, according to the file system.  This will prevent
browsers from caching them when they change.

TODO: for performance, in the future we should find a nice way to cache
the mtime check so that it only happens once per server restart.