Show 'New game' as title of the tab for any game without a sentence (closes #883).
Previously, only games in 'create' state were given the 'New game' title,
but this bug has "reminded" us, that games in 'create' state may also be
canceled, resulting in a game where sentence is not set, and the state
is 'canceled'.

So now if a game doesn't have a sentence set, the tab title will default
to 'New game', regardless of what state the game is actually in.