Animate continue button and next game info (closes #886).
The animation is a bit different from the one on the video.
The main difference is that the continue button doesn't pop
up until the level bar animation has finished (in the video,
the continue button pops up while the level bar animation is
still running).
This was done for simplicity - to change as little code as possible.
If I wanted to make it more similar to the video, I'd have to
rewrite a couple of times more code than this way.