Implement anonymous views for 'create' and 'invitation' states (closes #890).
I ran into a problem with the jQuery query plugin, because the query plugin
parses the url when the page is loaded and caches the results (it never parses
the url again), which doesn't work properly with the HTML5 History API where
URL can change without page being reloaded.

This made the query plugin unusable for the anonymous views where the 'anonymous'
query parameter needs to be checked, so I implemented a very simple replacement
for the query plugin, the $.cardstories.url_param function.

Currently the anonymous 'create' state doesn't differ from a 'create' state where
a player is actively participating in the game. We might want to change it in the future,
so that it will be easier for the anonymous player to know he is not part of the game.
I didn't bother with this yet, because a player who wants to join a full game
first sees the 'game is full' dialog, which might be sufficient.