Fix a bug where player's level score left is sometimes NaN (closes #887).
This also fixes some issues where the level score left would blink
before the final score was shown.

I thought about doing a test for this, but since it would have to look at the
DOM at certain exact point *during* the animation, the test would be hard to
write and would probably require me to put some hooks into the function to be
able to test it properly.
Since this was a rather trivial (and harmless) bug, I figured it wasn't
worth going through the trouble.