Cache file mtime values inside append_mtime() (fixes #884)
This uses the django cache engine to store file mtime values for 15
minutes (by default, timeout configurable in settings.py) in memory.
This prevents the main process having to access the filesystem multiple
times for every request.

An alternative would have been to cache just the template snippet where
the append_mtime tags are set, but this seemed to be a more transparent
(thus less error-prone) way to do it.