Implements score legend fly-in (fixes #870)
This implements the GAME SCORE characters flying in from the left.  This
was done dynamically, so that the phrase can be altered at any time with
no ill effects.

It took *a lot* of trial and error to get the formula for the parabola
right, but it'll pay off in flexibility for the future.