Implement levelup stars/fireworks animations (closes #897).
This implements the first part of the levelup animations:

- the win/lose banner morphing into the "Level up!" banner;
- the stars sprite with stars comming in and then blinking in a loop;
- the fireworks going off in a loop.

The loop stops whenever the active element gains the class cardstories_results_closed,
which is applied to it when clicking the continue button or closing the results
box with the little "X" button.

The level progress bar animations were made so that they first go from 0 to 100% for
each level the player has advanced (taking into account that it's possible to jump
ahead more than one level at once), and then showing the remaining progress
towards the next level.

The tests are not very extensive because most of the code is animations which
are hard or impossible to test well.