Implement earned cards backend (fixes #896)
This implements the mechanism through which a player earns new cards as
levels are gained.  The basic idea is simple: for every level earned,
the player gets a new card, chosen randomly from the cards above number
36.

In order to achieve this, the biggest change was that the games.cards
field no longer stores which cards haven't been dealt yet. It now stores
the exact opposite: which cards have already been dealt.  This was done
to accomodate the fact that each player now has a different set of
possible cards to be dealt: the base set of 36, plus an unknown number
of earned cards. Most of the rest of the changes stem off from this
change in architecture.

It is also worth nothing that the game master's initial deck is no
longer generated client-side.  When a game is created server-side, the
game master hand is immediately dealt and communicated to the client.

For the sake of simplicity, I decided to store earned cards using the
same technique as the rest of the card fields: a TEXT field with a
char for each card.  It complicates management later on, but the code is
easier to read and also gives better performance.