Level up animations: "You earned a new card" (fixes #900)
Displays the new card a player won, if any.

I've decided not to implement the display of multiple cards, for two
reasons.  First, the text is currently static, and it only refers to one
card.  Second, it is technically possible to earn 3 or 4 cards at a time
(that's what's going to happen the first time someone levels up after
the migration), and such behavior would be undefined as well.