Make tabs aware of tables (closes #804).
Tabs are now highlighted if the player is the owner of the next
game in the table, or if the next game in the table is ready
to be played.

The postprocess plugin hook was used to add next_owner_id and
next_player_id info into the response of the 'tabs' type state
request.

Because I needed to access the player_id in the postprocess hook,
I changed the signature of the postprocess function to take a second
parameter, which is the original request.
I believe that even in the postprocess step, having access to the
request is a completely valid expectation.

There are currently two problems with tabs highlighting when it
comes to tables:

- Because a new game of the table opens in a new tab,
  the tab of the previous game always remains highlighted
  (since the next lame is obviously ready).
- If a table is destroyed (such as after a server restart),
  each player is marked as the next owner, which means that
  after a server restart, all tabs containing games in the
  'complete' state will be highlighted.

The first problem should be solved by one of the future tasks,
where the next game of the table will open in the same tab as
the previous game (or will at least create that illusion by
automatically closing the previoius game tab), so that shouldn't be
a problem.

The second problem is trickier and I don't have an answer yet.