Open next table game in the series in the same tab (closes #800).
This doesn't really open the game in the *same* tab (I evaulated the options
to do that, but they would require much more work, and would have to rewrite
some parts of the system), instead it only closes the tab with the previous
game before loading the next one.

In order to do this I had to add a function to remove a tab by game_id programmatically
to the tabs plugin. Since I didn't like the idea of the table plugin calling
tabs plugin's functions, I instead change the table's load_next_game_when_ready
function to on_next_game_ready, which delegates the task of reloading the game
back to the core game (it simply invokes the passed in callback when the
next game is ready and doesn't do any reloading itself).

I took the opportunity to slightly simplify the table plugin by removing
the game2table indirection since we only ever deal with one table at a time.
I also did some JSLinting and added some missing semicolons.