Display player's name and level on the board (closes #910).
There is no styling yet, that will be done as a separate task.

I pulled the level caluculating code out from game.py into a separate
module to be able to use it from both game.py, and service.py.
It is used in service.py to add level information to the player_info
type call.

The updating of the display is done inside the $.cardstories.update_player_info
function, so that it is always in sync with the player_info as known to
the client.