Fix error propagation from chat (fixes #917)
When the new user-friendly oops message was implemented, the chat plugin
was not updated to take proper advantage of it.  As a consequence,
poll_plugin errors would not contain the necessary player_id to restart
said poll after display of the warning modal.

Ultimately, this which would cause an infinite loop of exceptions
server-side, since the player_id is necessary for authentication.

Notes:

* There were a few places beyond the actual client-side chat plugin
  where the player_id propagation was missing.  They were fixed.

* So as not to mask real errors behind an avalanche of
  player_id=undefined server-side exceptions, the poll_plugin is not
  restarted after an oops if player_id is not set.