Implement service tab_added/removed callbacks (closes #914).
This implements two new service "notifications".

When a tab is associated with a player in the tabs table, the listeners
are notified with a 'tab_added' type notification.
Similarly, when a tab is removed, they are notified with a 'tab_removed'
type notification.

I changed the tabInteraction to return an additional value which is
a boolean indicating whether the current game was added to the tabs
table (as opposed to not being added because it was already associated
with the player in the tabs table before this request).
I did this rather than calling service.notify directly inside the tabsInteraction
method because I figured it wouldn't be a good idea to fire notifications
while a DB transaction hasn't even finished yet.