Hook the table plugin into tab notifications to implement (closes #918).
This wires together the final pieces to implement story #787 (Detect when
the next author is not participating to a table anymore).

Since the table plugin was already reacting to player disconnect events
emitted by the activity plugin, adding a hook for a closed tab event
wasn't very hard as it reuses most of the code for handling disconnects.

I had some more trouble with twisted tests, though. To make the tests
work, I removed the helper methods start_table_poll and check_poll
that used mocks and instead used 'raw' poll mechanism. The advantage
of this is that it is possible to "block" the test waiting for the
poll to fire by yeilding to the poll. The problem I was seeing with
the mocked version was race conditions when checking whether the poll
has returned at a certain point in the test. Explicitly blocking until
the poll returns solves this issue, and makes the flow of the test
simpler to follow, IMHO.