Refactor the tab related service methods (closes #919).
The metods for opening and closing tabs went through a minor refactoring
so that they make more sense and are a bit easier to follow.

There is now also a service.get_opened_tabs(player_id) method which
returns a list of all opened tabs by a player, and doesn't do any other
magic under the hood.

The method that was called service.get_tab_game_ids previously and would
magically add the current game to opened tabs was renamed to
service.get_opened_tabs_from_args, which is a poor name and I would appreciate
any better suggestions.

Another method with a bad name is service.remove_tab, becaues it is easily confused
with service.close_tab. The difference between these to is that close_tab is just
a function that takes a player_id and a game_id and does what it says.
The remove_tab method on the other hand is an HTTP endpoint that the
JS client hits when the player closes the tab.
Again, suggestions for better names would be much appreciated (I am bad at this).

I also renamed the 'tab_added' and 'tab_removed' notifications to
'tab_opened' and 'tab_closed', which I think sounds better and makes more sense.