Better synchronize sounds on results page (refs #922).
I made the score_bell sound start playing a bit before the star
hits the end of the level bar, so that when the sound plays it isn't
too late already.

I also made the pinball sound start playing before the score
"slot machine" animation starts.
The pinball sound then keeps playing during the level bar animations
and is stopped when the level bar animations stop.
This probably isn't ideal (the slot machine animation is so
nice that it would deserve its own sound), but I tried a couple
of different combinations and this still seemed like the best
compromise.