Preload sound files via SoundManager's autoLoad option (close #920).
This creates all sound objects with the autoLoad option set to true.
This means that audio files will start loading as soon as the game
is loaded.

This creates a HTTP request for every sound at startup, but I think
we might be able to get away with this since we don't use too many
sound files, and all of them are quite small.

If it turns out this doesn't work well, sound manager offers options
for more fine-tuned preloading. We could probably even hook that into
the preloading screen for ultimate preload experience, but that would
require some more work.