Fix nasty bug when closing a tab in Firefox (closes #929).
The $.cardstories.send function was used from the tabs plugin,
and it wasn't called with the correct number of parameters
(I guess we forgot to update it when we were changing the
$.cardstories.send function).

The most problematic parameter that was missing was a callback
function, which $.cardstories.send invokes in a setTimeout.
Firefox seems to be mind this a lot since it throws an error
and stops execution, while most other browsers quietly ignore
this.

This commit fixes the tabs plugin to call send with the correct
number of arguments, while also making sure that send doesn't
try to invoke an undefined function in a setTimeout.