Skin the player level info display (closes #912).
Since there isn't a lot of place for the player's name,
the name will get truncated to 15 characters max.

There is a lot of duplication of the HTML structure,
but that's how things are currently set up and I still find it
cleaner (and easier) to duplicate the HTML than to dynamically
generate it with JS from a template.