Stop queues when changing tabs (refs #928, #927)
So that animations and sounds stop playing when the user changes tabs
(specifically level-up sounds when changing from a game in a complete
tab to another one), all queues must be stopped.  To do so, we must keep
track of queues: I chose to use the $(root) container as a global queue
element, in addition to a $(root).data('queues') array with the names of
the queues in use.

Most of the changes here involve refactoring the calls to .queue() to
accomodate the above.

However, this changeset also removes the debug code introduced in #927,
as I was finally able to reproduce the problem: it turns out that when
you open a tab of a previously completed game, the players_info[] array
will not contain the id of the next player.  I added some "duct tape"
so the error does not come up.  However, this may not be desired
behavior, and needs to be discussed.