Fix a bug when changing mind on the card (closes #936).
I introduced this bug while working on anonymous views for
the create steps. I had to change the way the closing of the
dock worked in the create_pick_card case, but accidently also
changed the behaviour for the case where the author changes his mind,
which should remain intact.

This commit fixes my mistake and adds a test case.