Fixes #941 - Bots don't pick a card - exceptions.KeyError: 39
The bots were still using NCARDS as the maximum number of card_id, not
taking into account the cards that the players can earn (NCARDS_EARNED).
When a player happened to chose a card he earned, the bot hadn't set an
initial value for that card's score, and was thus crashing.