Implement events log table (closes #938).
The table structure is basically the same as described in #65,
except that there is no field that indicates whether the event
belongs to the owner or the player because that should be evident
from the event type (author events don't overlap with player events).

I decided to store event types as integers and not as strings, which would
have been a bit more user friendly, but I decided against it as the event_logs
table has the potential to be quite large and storing event type as an integer
should be both more efficient and take up less space.

The event logging slows down the test_game.py tests quite a bit.
On my machine the tests run in 3.5 seconds with logging disabled and
in about 5 seconds with logging enabled. That looks like a big difference,
I wasn't really expecting it to be so noticeable.

The test_event_logs.py test is rather awkward because the logging is totally
async, so I came up with a hackish way of testing the events serially.