Implement functions to query/aggregate event logs (closes #957).
These functions are to be used for composing the daily emails.

They are split across two modules, some of them reside in the
cardstories.event_logs module itself, while others are part
of the mailing.aggregate module.

To be honest I'm not entirely sure why I split them like this.
The idea was the the more generic functions that only query
the event_logs table should be part of the event_logs module,
while those more specific to emails would be inside mailing.aggregate.
In the end I'm not really sure whether the event_logs function turned
out to be as generic as I imagined.

The idea is that for mail sending we will be using plain sync
python functions, not twisted's async stuff. This is to keep things
simple for now.