Fix the issue with tab not closing for the next GM of the table (closes #939).
The problem was with the poll that was running when the on_next_game_ready callback
was invoked. The callback first closes the current tab, which is an async operation.
When the tab is closed, the table plugin sees this as the next owner quitting the game,
which causes the table plugin to assign a new owner, and trigger the on_next_owner_change
event. This makes the poll return and issue another poll request with the "old" game_id.
The service sees a new poll request from a player that already closed the tab of that game,
so it thinks the owner is reopening the tab, and so adds the old game back to the tabs table,
which then reappears as an open tab on the player's screen.

This took me a long while to understand. The solution is to discards the running poll
prior to issuing the close tab request.

I also noticed there was a redundant call to on_next_game_ready, which I took
the opportunity to remove.