Fix invitation_pick animations (closes #959).
This one was caused by a line that slipped through unchanged through
refactoring. The fix is a one-liner, although it took me a while to find
the culprit, even after doing a git bisect.

There's no test for this one as I think it would be too much effort for
a bug that was essentially caused by a bad function call - and as such
isn't likely to happen again.