Implement and tie together missing pieces for the daily email feature (closes #964).
This implements fetching players from the django db, and iterating over them,
preparing data to be used in the email via the mailing.aggregate module functions.

The get_player_name in the mailing.aggregate module is a bit weird in that it
doesn't lookup the player id in the django database to get the correspoding name,
but instead expects to be "seeded" with info of all players in the db.
It then creates and stores a dict of player_id to player names in memory,
so that player name lookups are plain in memory dictionary lookups.

This was a bit of a (premature) optimization on my part, as I was initially thinking
to be fetching django data via HTTP requests, which would have been much heavier
than direct db lookups, and the optimization might make more sense then.
But the idea stuck, and since it actually makes some tests simpler,
I decided to keep it.

The mailing.send module will need to be updated to work with the new mailing.loop module.

The ability to opt-out of receiving the daily emails still needs to be implemented
on the django side and this code will have to be updated to respect it once it's done
(but the changes should be minor).