Fix KeyError bug in table.py (closes #950).
The problem was with the code for the closed tab handler in the table plugin.
The code blindly assumed a table for game that was being closed existed,
which isn't the case for games that were created before the last server restart.

I believe the best fix (which would also solve some of the other problems of tables)
would be to make tables persistent somehow, so that not everything is lost after
a server restart. That would make a lot of things easier, but would require a
major refactoring and isn't something we could afford right now,
so I only added a check for table existing in game2table dict before trying
to call the on_tab_closed handler on it.