Show game info on Facebook feed (refs #965 and #966)
This implements the feature specified on #932 where when the game master
clicks "like" on the invitation box, the corresponding item on his
timeline/feed should contain a link to the current game, as well as the
game's story.

Showing the copy-pastable URL to the current game on the invitation box
was also implemented here.  I tried to get it to copy-to-clipboard on
click, but apparently that's a security liability and therefore not
practical in any sane cross-browser way.

Notes:

* The feature where the user could enter a list of emails to invite was
  completely removed client-side (but not server-side).

* Since we're including the Facebook JS SDK anyway, the top Like button
  now uses it, too, instead of the iframe.

* The Facebook App ID must be set correctly in settings.py or
  local_settings.py, otherwise the Like buttons won't work (they did
  before this, but the app id's were hardcoded).

* The Facebook SDK is now loaded by the Django template, as opposed to
  being loaded at runtime when the invitation box is opened.  This
  should give the impression that it's speedier.

* Django templates were a bit rearranged where the Facebook and G+
  SDKs were concerned: they're now included by default in the base
  template (since they're always used).