Fix 'NoneType' object is unsubscriptable errors (closes #949).
The problem was that when the game is destroyed, the poll returns with None.
This was implemented intentionally, though I am not sure what the reason
behind it was.

It was tripping up the tab polling code because if one of the games in the
tabs was cancelled (when game is cancelled, destroy is called on it),
poll returned with None and the code that was looking for the 'modified'
argument errored out.

I solved this by touching the game with a type='cancel' touch before destroying
it in the cancel method. This makes any polls return gracefully before the game
is destroyed. When a user has a canceled game opened, it now correctly detects
it changed state and shows our "Game has been canceled" popup.
Without this change, the client code was erroring out because 'modified' wasn't
there.

Another possible source of game.destroy() calls is when shutting down the service.
All games are destroyed then. The errors were harmless, but I added a check
for the returned poll arguments being equal to None, before trying to search
for 'modified', to silence the errors.