Don't stop listening for events when AssertError is raised (closes #974).
The reason why the game was in a broken state some days ago was that one of
the notification callbacks modified the 'modified' flag of the game it listened
for events to, which is forbidden. There is an assertion in place to enforce that.

Unfortunately, if the assertion was raised, the service didn't re-attach the
listen callback to the game, causing the listen "loop" to break. This led to the
service not being notified when that game was destroyed, so that it failed to
be deleted from the service.games hash, causing the avalanche of errors that
we saw in production.

While it is essential that we know when one of the notification callbacks modifies
the game even though it shouldn't, the service should try to go on (while logging the error)
and not just get itself into a broken state. That's what this commit is trying
to achieve.