Don't break on nonexisting game_ids (closes #977).
The tabs code was broken when a request with a nonexisting game_id
was made to the service. Instead of showing the 'Game does not exist'
dialog to the player, the game would error out, and what's worse,
the game would keep erroring for that player even if he tried to visit
an existing game (or create a new one), because the faulty game_id
was already stored and associated with the player in the tabs table.

This commit prevents nonexisting game_ids from getting into the tabs table.

It also includes a migration that deletes all rows from the tabs table
that hold a game_id that refers to a game that doesn't exist.