Don't send out blank emails (closes #981).
If the player was active not long ago before doing the email sending,
and there is no new info for him to be emailed about, make sure
that the email isn't sent.

We already have a 'should_send_email' function so it would be nice if
this additional check could be added to it, but the idea was if we already know
we don't want to send the email from other conditions, we dan't want to
go fetching everything about the player from the db, and this additinal
check can only be made after everything has already been fetched from the db,
so I kept it separate from 'should_send_email'.