Make it possible to only send emails to specified recipients (closes #982).
The loop function takes an additional optional email_list parameter,
which when present, acts as a filter to only allow sending email
to emails present in the list.

When invoking the loop.py script from the shell directly, path to a file
containing newline separated emails can be given as the last argument.