Fix invite dialog layout in FF on OSX (closes #979).
It took me a while to realize why the three inner elements didn't want
to line up (the last one was pushed down to the next row), even though
the widths of the inner elements seemed to have been correctly set to
not exceed the width of the parent.
While I could have simply reduced the widths of the inner elements,
I wanted to understand what was happening.

It turned out the problem was that the three divs are displayed as
'inline-block', which preserves spaces between elements, if spaces are
present in the HTML source. So between the three elements there were
spaces, taking up extra horizontal space.

I found this website which lists a couple of tricks about how to get rid
of the whitespace between inline-block elements:

http://css-tricks.com/fighting-the-space-between-inline-block-elements/

I chose the most straightforward method - removing whitespace between
elements in the HTML source itself.