Fix some service.game_notify concurrency problems (closes #978).
The assertion that was asserting that no listeners modifies the game
during service notifications was removed, because it didn't take
asynchronous execution into account.

Also due to the asynchronous nature of notifications, it was possible that
some events were silently 'dropped' and not dispatched to the listeners.
This has been corrected, and a test case added.

For more background and a more detailed discussion see:

http://tickets.farsides.com/issues/978