Make table discard game that takes to long to be created (closes #985).
When a new game from the table is ready, the table starts a timer
which fires after a predefined delay (currently 1 minute) and checks
whether the game was moved into 'invitation' state by that time yet.
If it hasn't, it "rolls back" the state of the table, discarding the game,
pointing back to the previous game as the current game of the table,
while choosing a new next_owner.

There is still one problem that crossed my mind, which is that if
a discarded game ever makes it into 'invitation' state, it should
be completely dettached from the table, and moved into a brand new
table; or the table state could get seriously corrupt if two games
were played and tracked under the same table simoultaneously.